requirements.txt
azure-ai-translation-text
deepl
dotenv
discord.py